feat(crypto): 가상화폐 모의투자 기능 entity 및 repository 개발-#55 (#57)
- 거래를 하나의 event로, DB table의 append-only 방식
- offset 방식은 건너뛸 데이터까지 DB가 순차적으로 모두 읽고 버려야 하는 구조라, WHERE 구문으로 처리 가능한 ProcessorOffset Entity 추가

+++ tradelink/src/main/java/site/tradelink/tradelink/cryptocurrency/sse/SseEmitterManager.java
@@ -11,6 +11,7 @@ import site.tradelink.tradelink.cryptocurrency.dto.StockPriceSummaryDto;
 import site.tradelink.tradelink.cryptocurrency.dto.TradeLogDto;
 import site.tradelink.tradelink.cryptocurrency.inMemory.OrderBookCache;
 import site.tradelink.tradelink.cryptocurrency.inMemory.StockPriceCache;
+import site.tradelink.tradelink.cryptocurrency.sse.dto.MyOrderDto;
 import site.tradelink.tradelink.cryptocurrency.sse.dto.SseEvent;
 
 import java.time.LocalDateTime;
@@ -126,7 +127,7 @@ public class SseEmitterManager {
 
     // MatchingEngine이 직접 호출, 해당 멤버만
     public void pushMyOrder(Long memberSeq, String ticker, long price,
-                            long quantity, String side, String status,
+                            double quantity, String side, String status,
                             LocalDateTime at) {
 
         MyOrderDto dto = new MyOrderDto(ticker, price, quantity, side, status, at);
